+++ lib/marin/src/marin/defaults/step.py
@@ -1,61 +1,40 @@
-# Copyright The Marin Authors
-# SPDX-License-Identifier: Apache-2.0
 import dataclasses
 import os
-from collections.abc import Sequence
-from dataclasses import dataclass
 from datetime import timedelta
 from functools import lru_cache
-from typing import Any
+from typing import Sequence
 
 import jmp
-from fray import ResourceConfig
+
 from haliax.partitioning import ResourceAxis
 from haliax.quantization import QuantizationConfig
-from levanter.adaptor import AdaptorConfig, LoraAdaptorConfig, NoAdaptorConfig
-from levanter.callbacks.profiler import ProfilerConfig
-from levanter.callbacks.watch import WatchConfig
+from levanter.adaptor import AdaptorConfig, NoAdaptorConfig, LoraAdaptorConfig
 from levanter.checkpoint import CheckpointerConfig
-from levanter.data.text import (
-    DEFAULT_LM_DATA_SHUFFLE,
-    LmDatasetFormatBase,
-    LMMixtureDatasetConfig,
-    PreferenceLmDataConfig,
-    TextLmDatasetFormat,
-)
-from levanter.dpo import ReferenceEvalCacheConfig
+from levanter.data.text import LMMixtureDatasetConfig, PreferenceLmDataConfig, \
+  DEFAULT_LM_DATA_SHUFFLE
 from levanter.eval_harness import LmEvalHarnessConfig
-from levanter.main.train_dpo import DpoReferenceConfig, SeparateReferenceConfig, TrainDpoConfig
+from levanter.main.train_dpo import SeparateReferenceConfig, TrainDpoConfig
 from levanter.main.train_lm import TrainLmConfig
 from levanter.models.llama import LlamaConfig
-from levanter.models.lm_model import LmConfig
-from levanter.optim import AdamConfig, EmaModelAveragingConfig, OptimizerConfig
-from levanter.schedule import BatchSchedule, IntSchedule
-from levanter.tracker.wandb import WandbConfig, truncate_wandb_run_name
-from levanter.trainer import TrainerConfig
-from levanter.utils import fsspec_utils
-from levanter.utils.mesh import MeshConfig
 
-from marin.datakit.download.huggingface import DownloadConfig, download_hf
+from levanter.models.lm_model import LmConfig
+from levanter.optim import EmaModelAveragingConfig, AdamConfig
+from levanter.schedule import BatchSchedule
+from levanter.tracker.wandb import truncate_wandb_run_name, WandbConfig
+from levanter.trainer import TrainerConfig
+from levanter.utils.mesh import MeshConfig
 from marin.datakit.download.uncheatable_eval import make_uncheatable_eval_step
-from marin.evaluation.evaluation_config import EvalTaskConfig, convert_to_levanter_task_config
+from marin.defaults.config import SimpleTrainConfig, SimpleSFTConfig, SimpleDPOConfig
+from marin.defaults.tokenize import default_tokenize
+from marin.evaluation.evaluation_config import EvalTaskConfig, \
+  convert_to_levanter_task_config
 from marin.execution.executor import unwrap_versioned_value
-from marin.execution.remote import remote
-from marin.execution.types import ExecutorStep, InputName, VersionedValue, ensure_versioned, this_output_path, versioned
-from marin.processing.tokenize import (
-    TokenizeConfig,
-    TokenizerStep,
-    add_validation_sets_to_mixture,
-    lm_data_config,
-    lm_mixture_data_config,
-)
-from marin.processing.tokenize.tokenize import HfDatasetSpec, HfTokenizeConfig, TokenizeConfigBase
-from marin.training.training import (
-    TrainDpoOnPodConfig,
-    TrainLmOnPodConfig,
-    run_levanter_train_dpo,
-    run_levanter_train_lm,
-)
+from marin.execution.types import InputName, ExecutorStep, this_output_path, versioned
+from marin.processing.tokenize import TokenizerStep, lm_mixture_data_config, \
+  lm_data_config, add_validation_sets_to_mixture, TokenizeConfig
+from marin.processing.tokenize.tokenize import TokenizeConfigBase, HfTokenizeConfig
+from marin.training.training import TrainLmOnPodConfig, run_levanter_train_lm, \
+  TrainDpoOnPodConfig, run_levanter_train_dpo
 
 
 def _resolve_hf_export_steps(steps_per_hf_export: int | None, steps_per_export: int | None) -> int | None:
@@ -95,102 +74,6 @@ def default_validation_sets(tokenizer: str, base_path: str = "tokenized/") -> di
     validation_sets = dict(paloma_tokenized(base_path=base_path, tokenizer=tokenizer))
     validation_sets.update(uncheatable_eval_tokenized(base_path=base_path, tokenizer=tokenizer))
     return validation_sets
-
-
-@dataclasses.dataclass(frozen=True)
-class SimpleTrainConfig:
-    resources: ResourceConfig
-    train_batch_size: int | IntSchedule
-    """
-    The batch size for training. If an IntSchedule is provided, the batch size will be
-    varied according to the schedule.
-    """
-    num_train_steps: int
-    learning_rate: float
-    train_seq_len: int | None = None
-    data_seed: int | None = None
-    weight_decay: float | None = None
-    beta1: float | None = None
-    beta2: float | None = None
-    epsilon: float | None = None
-    max_grad_norm: float | None = None
-    warmup: float | None = None
-    decay: float | None = None
-    rewarmup: float | None = None
-    """
-    The rewarmup parameter is used to re-warmup the learning rate after a decay cycles
-    """
-    lr_schedule: str | None = None
-    min_lr_ratio: float | None = None
-    cycle_length: int | list[int] | None = None
-    z_loss_weight: float | None = None
-    ema_beta: float | None = None
-    """exponential moving average beta"""
-    skip_bad_steps: bool = False
-    """If True, skips steps where the loss or grad is significantly higher than the historical mean."""
-
-    steps_per_eval: int | None = None
-    """how often to run validation losses"""
-    steps_per_export: int | None = None
-    """How often to keep a permanent checkpoint. None (default) keeps only the final
-    checkpoint; rolling temporary checkpoints are still written for resumption."""
-    steps_per_task_eval: int | None = None
-    """how often to run task evaluations"""
-    steps_per_hf_export: int | None = None
-    """None means match steps_per_export, -1 disables"""
-    hf_generation_eos_token_ids: list[int] | None = None
-    """EOS token IDs to write to generation_config.json. None means no generation config."""
-    per_device_parallelism: int = -1
-    """How many examples to process in parallel on each device. -1 (default) means
-    train_batch_size/num_devices (no gradient accumulation). Set to a positive value
-    to enable gradient accumulation."""
-    per_device_eval_parallelism: int | None = None
-    """Number of examples to evaluate in parallel on each device"""
-    max_eval_batches: int | None = None
-    """Maximum number of batches to evaluate on. None means all batches"""
-
-    initialize_from_checkpoint_path: str | None = None
-    """If set, the training will resume from the checkpoint at this path. Otherwise, training will start from scratch."""
-    initialize_from_hf: str | None = None
-    """If set, the training will start from the hf model at this path. Otherwise, training will start from scratch."""
-    reset_data_loader_on_init: bool = True
-    """Pairs with initialize_from_checkpoint_path. If True, initialize_from_checkpoint_path will reset the data loader
-    so that it starts from step 0. Otherwise, it will resume from the step in the checkpoint."""
-
-    allow_partial_checkpoint: bool = False
-    """
-    Allow loading partial checkpoints. This is useful for converting training to EMA, e.g.
-    """
-
-    int8: bool = False
-    """Int8 (quantized) training in Levanter."""
-
-    pad_tokenizer_to_match_model: bool = False
-    """If True, pad the tokenizer's vocab to match the model's vocab size by adding dummy tokens.
-    Useful when the model checkpoint has a larger vocab than the tokenizer (e.g., Qwen models
-    pad their vocab to be divisible by 4 for TPU efficiency)."""
-
-    optimizer_config: OptimizerConfig | None = None
-    """Optimizer configuration to use. If not set, Adam will be used."""
-
-    watch: WatchConfig = dataclasses.field(default_factory=WatchConfig)
-    """Config for watching gradients, parameters, etc. Default is to log norms of gradients and parameters."""
-
-    profiler: ProfilerConfig = dataclasses.field(default_factory=ProfilerConfig)
-    """JAX profiler settings for training."""
-
-    explicit_mesh_axes: bool = False
-    """If True, build the device mesh with `AxisType.Explicit` axes.
-
-    Required for models that call `jax.sharding.reshard(..., PartitionSpec(...))`.
-    """
-
-    tensor_parallel_size: int = 1
-    """Size of the model (tensor parallel) axis. >1 shards model weights and activations
-    across multiple devices. Useful when batch_size < num_chips."""
-
-    env_vars: dict[str, str] | None = None
-    """Environment variables to pass to the training task."""
 
 
 # tasks to run (corresponding to lm_eval_harness tasks)
@@ -426,123 +309,6 @@ def default_train(
     )
 
 
-@dataclass(frozen=True)
-class SimpleSFTConfig:
-    """
-    A simplified configuration for Supervised Fine-Tuning (SFT) that works for both
-    single dataset and mixture training approaches.
-    """
-
-    # Hardware configuration
-    resources: ResourceConfig
-
-    # Core training parameters
-    train_batch_size: int | IntSchedule = 128
-    """
-    The batch size for training. If an IntSchedule is provided, the batch size will be
-    varied according to the schedule.
-    """
-    num_train_steps: int = 10000
-    """Number of training steps."""
-
-    learning_rate: float = 5e-6
-    """Learning rate for the optimizer."""
-
-    # Model configuration
-    tokenizer: str | None = None
-    """Tokenizer to use for training."""
-
-    initialize_from_hf: str | None = None
-    """HF model name or path to initialize from (e.g., 'meta-llama/Llama-3.1-8B').
-    Mutually exclusive with initialize_from_checkpoint_path."""
-
-    initialize_from_checkpoint_path: str | None = None
-    """Path to a levanter checkpoint to initialize from.
-    Mutually exclusive with initialize_from_hf."""
-
-    max_seq_len: int = 4096
-    """Maximum sequence length for training."""
-
-    # Optimizer parameters
-    weight_decay: float = 0.0
-    """Weight decay for the optimizer."""
-
-    beta1: float | None = None
-    """AdamW optimizer beta1."""
-
-    beta2: float | None = None
-    """AdamW optimizer beta2."""
-
-    warmup: float = 0.03
-    """Fraction of training steps to use for learning rate warmup."""
-
-    decay: float = 0.0
-    """Fraction of training steps to use for learning rate decay."""
-
-    lr_schedule: str = "linear"
-    """Learning rate schedule to use: 'linear', 'cosine', etc."""
-
-    min_lr_ratio: float = 0.0
-    """Minimum learning rate as a ratio of the base learning rate."""
-
-    max_grad_norm: float | None = None
-    """Maximum gradient norm for gradient clipping."""
-
-    # Checkpointing and evaluation
-    steps_per_eval: int = 1000
-    """How often to run validation losses."""
-
-    steps_per_checkpoint: int | None = None
-    """How often to keep a permanent checkpoint. None (default) keeps only the final
-    checkpoint; rolling temporary checkpoints are still written for resumption."""
-
-    steps_per_hf_export: int = 500
-    """How often to save HuggingFace checkpoints."""
-
-    hf_generation_eos_token_ids: list[int] | None = None
-    """EOS token IDs to write to generation_config.json. None means no generation config.
-    For chat models, include the turn-boundary token (e.g. [128001, 128009])."""
-
-    # Mixture-specific parameters
-    mixture_block_size: int = 2048
-    """Block size for dataset mixing (only used with mixture training)."""
-
-    stop_strategy: str = "restart"
-    """
-    Strategy for handling dataset completion (only used with mixture training).
-    Options: 'restart' or 'exit'.
-    """
-
-    # Other parameters
-    seed: int = 0
-    """Random seed for training."""
-
-    node_count: int = 1
-    """Number of TPU slices for training."""
-
-    int8: bool = False
-    """Int8 (quantized) training in Levanter."""
-
-    pad_tokenizer_to_match_model: bool = False
-    """If True, pad the tokenizer's vocab to match the model's vocab size by adding dummy tokens.
-    Useful when the model checkpoint has a larger vocab than the tokenizer (e.g., Qwen models
-    pad their vocab to be divisible by 4 for TPU efficiency)."""
-
-    z_loss_weight: float = 0.0
-
-    per_device_parallelism: int = -1
-    """How many examples to process in parallel on each device. -1 (default) means
-    train_batch_size/num_devices (no gradient accumulation). Set to a positive value
-    to enable gradient accumulation. For example, with 8 devices, batch_size=32, and
-    per_device_parallelism=1, you get gradient accumulation of 4."""
-
-    reinit_tokens: list[str] | bool = False
-    """
-    if set, will reinitialize the embeddings for the given tokens. If True, will reinitialize the default tokens
-    for llama3's tokenizer
-    """
-
-
 def default_sft(
     name: str,
     tokenized: InputName | ExecutorStep | LMMixtureDatasetConfig,
@@ -615,77 +381,6 @@ def default_sft(
         use_default_validation=False,
         adapter=adapter,
     )
-
-
-@dataclass(frozen=True)
-class SimpleDPOConfig:
-    """
-    A simplified configuration for Direct Preference Optimization (DPO).
-    """
-
-    resources: ResourceConfig
-
-    train_batch_size: int | IntSchedule = 128
-    num_train_steps: int | None = None
-    num_epochs: float = 1.0
-    """Approximate number of passes over the DPO train set when num_train_steps is unset."""
-    learning_rate: float = 1e-6
-    wandb_project: str | None = None
-
-    tokenizer: str | None = None
-    model_name_or_path: str | None = None
-    initialize_from_checkpoint_path: str | None = None
-
-    adapter: AdaptorConfig = dataclasses.field(default_factory=NoAdaptorConfig)
-    reference: DpoReferenceConfig = dataclasses.field(default_factory=SeparateReferenceConfig)
-    reference_model_path: str | None = None
-    reference_is_hf: bool = True
-    beta: float = 0.1
-    validation_split_fraction: float | None = 0.1
-    reference_eval_cache: ReferenceEvalCacheConfig = dataclasses.field(
-        default_factory=lambda: ReferenceEvalCacheConfig(mode="build_or_load")
-    )
-
-    train_seq_len: int | None = None
-    max_seq_len: int = 4096
-
-    weight_decay: float = 0.0
-    warmup: float = 0.0
-    cooldown: float | None = None
-    lr_schedule: str = "linear"
-    min_lr_ratio: float = 0.0
-    max_grad_norm: float | None = 1
-
-    steps_per_eval: int | None = None
-    """None auto-schedules validation five times: before training, three interior points, and at the end."""
-    steps_per_checkpoint: int | None = None
-    """How often to keep a permanent checkpoint. None (default) keeps only the final
-    checkpoint; rolling temporary checkpoints are still written for resumption."""
-    steps_per_hf_export: int = 500
-    hf_save_dtype: str | None = None
-    hf_generation_eos_token_ids: list[int] | None = None
-    """EOS token IDs to write to generation_config.json. None means no generation config.
-    For chat models, include the turn-boundary token (e.g. [128001, 128009])."""
-
-    per_device_eval_parallelism: int = -1
-
-    seed: int = 0
-    initialize_from_hf: bool | None = None
-
-    profiler: ProfilerConfig = dataclasses.field(default_factory=ProfilerConfig)
-
-    allow_partial_checkpoint: bool = False
-    int8: bool = False
-
-    def __post_init__(self):
-        if self.num_train_steps is not None and self.num_train_steps <= 0:
-            raise ValueError(f"num_train_steps must be positive, got {self.num_train_steps}")
-        if self.num_epochs <= 0:
-            raise ValueError(f"num_epochs must be positive, got {self.num_epochs}")
-        if self.steps_per_eval is not None and self.steps_per_eval <= 0:
-            raise ValueError(f"steps_per_eval must be positive, got {self.steps_per_eval}")
-        if self.steps_per_checkpoint is not None and self.steps_per_checkpoint <= 0:
-            raise ValueError(f"steps_per_checkpoint must be positive, got {self.steps_per_checkpoint}")
 
 
 def default_dpo(
@@ -961,215 +656,3 @@ ACTIVE_DATASETS = [
     "arxiv_computer_science",
     "ao3_english",
 ]
-
-
-def compute_per_device_parallelism(
-    global_batch_size: int,
-    microbatch_size: int,
-    resources: ResourceConfig,
-) -> int:
-    """Compute per_device_parallelism for gradient accumulation.
-
-    Args:
-        global_batch_size: The effective batch size after gradient accumulation.
-        microbatch_size: The batch size that fits in memory (local batch size).
-        resources: The ResourceConfig specifying TPU/GPU resources.
-
-    Returns:
-        per_device_parallelism: Number of examples each device processes per forward/backward pass.
-
-    Example:
-        For v5p-8 (4 chips), global_batch_size=128, microbatch_size=8:
-        - per_device_parallelism = 8 / 4 = 2
-        - gradient_accumulation = 128 / 8 = 16 steps
-    """
-    num_devices = resources.chip_count()
-
-    if microbatch_size % num_devices != 0:
-        raise ValueError(f"microbatch_size ({microbatch_size}) must be divisible by " f"num_devices ({num_devices})")
-
-    if global_batch_size % microbatch_size != 0:
-        raise ValueError(
-            f"global_batch_size ({global_batch_size}) must be divisible by " f"microbatch_size ({microbatch_size})"
-        )
-
-    per_device_parallelism = microbatch_size // num_devices
-    grad_accum_steps = global_batch_size // microbatch_size
-
-    print(
-        f"Gradient accumulation config: "
-        f"global_batch={global_batch_size}, microbatch={microbatch_size}, "
-        f"num_devices={num_devices}, per_device_parallelism={per_device_parallelism}, "
-        f"grad_accum_steps={grad_accum_steps}"
-    )
-
-    return per_device_parallelism
-
-
-HF_BUCKET_URI_PREFIX = "hf://buckets/"
-HF_BUCKET_PATH_PREFIX = "buckets/"
-
-
-def _is_hf_bucket_path(path: str) -> bool:
-    return path.startswith(HF_BUCKET_URI_PREFIX) or path.startswith(HF_BUCKET_PATH_PREFIX)
-
-
-def _normalize_hf_bucket_path(path: str) -> str:
-    if path.startswith(HF_BUCKET_URI_PREFIX):
-        return path.removeprefix("hf://")
-    return path
-
-
-def default_download(
-    name: str,
-    hf_dataset_id: str,
-    revision: str | None = None,
-    override_output_path: str | None = None,
-    **kwargs: Any,
-) -> InputName:
-    """
-    Download a HuggingFace dataset and upload it to a specified path with default configuration.
-
-    Args:
-        name: The name of the Download step. It forms the basis of the output path
-            unless override_output_path is explicitly specified.
-        hf_dataset_id: Hugging Face source. Either `$ORG/$DATASET` on HF Hub or `hf://buckets/...`.
-        revision: The revision of the dataset to download for Hub datasets.
-            Optional for bucket paths.
-        override_output_path: Optional. The output path for the dataset.
-        **kwargs: Additional keyword arguments that are passed to the download config.
-
-    The final output data will reside in '{output_path}/{revision}'.
-    """
-
-    download_kwargs = dict(kwargs)
-    hf_repo_type_prefix = download_kwargs.pop("hf_repo_type_prefix", None)
-    if _is_hf_bucket_path(hf_dataset_id):
-        normalized_dataset_id = _normalize_hf_bucket_path(hf_dataset_id)
-        description = f"Download {hf_dataset_id}"
-        resolved_hf_repo_type_prefix = "" if hf_repo_type_prefix is None else hf_repo_type_prefix
-        resolved_revision = "main" if revision is None else revision
-    else:
-        if revision is None:
-            raise ValueError("revision is required for non-bucket Hugging Face dataset downloads.")
-        normalized_dataset_id = hf_dataset_id
-        description = f"Download {hf_dataset_id} revision {revision}"
-        resolved_hf_repo_type_prefix = "datasets" if hf_repo_type_prefix is None else hf_repo_type_prefix
-        resolved_revision = revision
-
-    step = ExecutorStep(
-        name=name,
-        description=description,
-        fn=download_hf,
-        config=DownloadConfig(
-            hf_dataset_id=normalized_dataset_id,
-            revision=resolved_revision,
-            gcs_output_path=this_output_path(),
-            wait_for_completion=True,
-            hf_repo_type_prefix=resolved_hf_repo_type_prefix,
-            **download_kwargs,
-        ),
-        override_output_path=override_output_path,
-    )
-
-    return step.as_input_name()
-
-
-def default_tokenize(
-    name: str,
-    dataset: InputName | ExecutorStep | str | HfDatasetSpec,
-    tokenizer: str,
-    format: LmDatasetFormatBase = TextLmDatasetFormat(),  # noqa
-    *,
-    sample_count: int | VersionedValue[int] | None = None,
-    is_validation: bool = False,
-    levanter_batch_size: int | None = None,
-    tags: Sequence[str] = (),
-    resources: ResourceConfig | None = None,
-    worker_resources: ResourceConfig | None = None,
-) -> ExecutorStep:
-    """
-    Tokenizes a dataset using the specified tokenizer and Levanter's tokenization infrastructure.
-
-    Args:
-        name: The name of the tokenized dataset. This is used to form the output path for the executor step.
-            `tokenized/` will be prepended to the name.
-        dataset:  The dataset to tokenize. This can be an InputName, ExecutorStep, a string as a
-            path to the dataset or a HuggingFace dataset ID, or ``HfDatasetSpec`` to specify a
-            dataset with a particular subset name.
-        tokenizer: string HuggingFace tokenizer name. Should be the same as you intend to use in the tokenizer
-            spec for the training run.
-        format: The format of the dataset. This is used to determine how to tokenize the data.
-
-            See [Levanter's documentation](https://levanter.readthedocs.io/en/latest/reference/Data-Formats/)
-            for more details.
-        sample_count: Optional limit on the number of samples to tokenize per shard. If ``None``, tokenize everything.
-        is_validation: Whether the dataset is a validation set. Doesn't do anything for HF datasets.
-        tags: Tags to attach to the Levanter dataset source for tagged evaluation.
-    Returns:
-        An ExecutorStep that represents the tokenized dataset.
-    """
-
-    # Common kwargs for config constructors
-    extra_kwargs: dict = {}
-    if worker_resources is not None:
-        extra_kwargs["worker_resources"] = worker_resources
-
-    # sniff out if it's a HuggingFace dataset
-    if isinstance(dataset, HfDatasetSpec):
-        config = HfTokenizeConfig(
-            id=dataset.id,
-            name=dataset.name,
-            cache_path=this_output_path(),
-            tokenizer=ensure_versioned(tokenizer),
-            format=format,
-            sample_count=ensure_versioned(sample_count) if sample_count is not None else None,
-            levanter_batch_size=levanter_batch_size,
-            tags=[*tags],
-            **extra_kwargs,
-        )
-    elif (
-        isinstance(dataset, str)
-        and not _is_hf_bucket_path(dataset)
-        and dataset.count("/") == 1
-        and not fsspec_utils.exists(dataset)
-    ):
-        config = HfTokenizeConfig(
-            id=dataset,
-            cache_path=this_output_path(),
-            tokenizer=ensure_versioned(tokenizer),
-            format=format,
-            sample_count=ensure_versioned(sample_count) if sample_count is not None else None,
-            levanter_batch_size=levanter_batch_size,
-            tags=[*tags],
-            **extra_kwargs,
-        )
-    else:
-        config = TokenizeConfig(
-            train_paths=[dataset] if not is_validation else [],
-            validation_paths=[dataset] if is_validation else [],
-            cache_path=this_output_path(),
-            tokenizer=ensure_versioned(tokenizer),
-            format=format,
-            sample_count=ensure_versioned(sample_count) if sample_count is not None else None,
-            levanter_batch_size=levanter_batch_size,
-            tags=[*tags],
-            **extra_kwargs,
-        )
-
-    return ExecutorStep(
-        name=os.path.join("tokenized", name),
-        description=f"Tokenize raw text using the {tokenizer} tokenizer.",
-        fn=remote(
-            tokenizer,
-            resources=resources or ResourceConfig.with_cpu(cpu=4, ram="16g", disk="10g"),
-            pip_dependency_groups=["cpu"],
-            env_vars={
-                "TRANSFORMERS_NO_TORCH": "1",
-                "TRANSFORMERS_NO_TORCHVISION": "1",
-                "USE_TORCH": "0",
-                "TORCH_DISABLE_GLOBAL_DEPS": "1",
-            },
-        ),
-        config=config,
-    )
